requirements.txt
requests>=2.28.0
pytz>=2022.1
google-cloud-storage>=2.7.0
google-cloud-logging>=3.2.0
python-dotenv>=0.19.0
dataclasses-json>=0.5.7